services/dashboard-api/requirements.txt
Flask>=2.3.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
